**Upgrading To Laravel 5.5**
Still Working on the Portfolio Part.

However, as New Laravel 5.5 is out. I am Eager to try that out.

So before Upgrading this Project to 5.5, I want to Commit and Push the Existing Changes, so that it will be easy to roll back to previous version if i get in to any problems. :)